refactor(save-cordova-xml): Rename to more descriptive update config util
