+++ src/utils/update-config.js
